Add readline tests for pdb completion
